# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/site/policies/{kind}": {
      "get": {
        "operationId": "SettingsController_policy",
        "parameters": [
          {
            "name": "kind",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/admin/settings": {
      "get": {
        "operationId": "SettingsController_getSite",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "SettingsController_updateSite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "contactPhone": {
                    "type": "string",
                    "maxLength": 40
                  },
                  "contactEmail": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "email",
                        "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                      },
                      {
                        "type": "string",
                        "const": ""
                      }
                    ]
                  },
                  "contactAddressFa": {
                    "type": "string",
                    "maxLength": 400
                  },
                  "contactAddressEn": {
                    "type": "string",
                    "maxLength": 400
                  },
                  "workingHoursFa": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "workingHoursEn": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "domain": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "reservationMinutes": {
                    "default": 15,
                    "type": "integer",
                    "minimum": 5,
                    "maximum": 60
                  },
                  "quoteValidityHoursDefault": {
                    "default": 24,
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 720
                  },
                  "taxRateBasisPoints": {
                    "default": null,
                    "anyOf": [
                      {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 5000
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "taxBase": {
                    "default": "ITEMS",
                    "type": "string",
                    "enum": [
                      "ITEMS",
                      "ITEMS_AND_SHIPPING"
                    ]
                  },
                  "manualBankTransferEnabled": {
                    "default": false,
                    "type": "boolean",
                    "const": false
                  },
                  "version": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  }
                },
                "required": [
                  "version"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/admin/policies": {
      "get": {
        "operationId": "SettingsController_policies",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "settings"
        ]
      },
      "post": {
        "operationId": "SettingsController_createPolicy",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "kind": {
                    "type": "string",
                    "enum": [
                      "TERMS",
                      "PRIVACY",
                      "RETURNS",
                      "SHIPPING",
                      "SOURCING",
                      "WARRANTY"
                    ]
                  },
                  "titleFa": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 200
                  },
                  "titleEn": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 200
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "bodyFa": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 50000
                  },
                  "bodyEn": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 50000
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "required": [
                  "kind",
                  "titleFa",
                  "bodyFa"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/admin/policies/{id}/publish": {
      "post": {
        "operationId": "SettingsController_publishPolicy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "settings"
        ]
      }
    },
...
    "/api/v1/admin/calendars/{id}": {
      "put": {
        "operationId": "SettingsController_updateCalendar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "settings"
        ]
      }
    },
...
  },
  "info": {
    "title": "HEDAX API",
    "description": "Auto parts store and custom sourcing platform — REST API v1. Money is always { currency, amountMinor: string }.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hedax_sid"
      }
    },
...
  }
}